teste

+++ Backend/src/controllers/controllerSupplier.ts
@@ -291,8 +291,12 @@ export const supplierDelete = async (req: Request, res: Response): Promise<void>
 
     // create resource
     const supplierDeleted = await prisma.supplier.delete({ where: { uuid: params.data.uuid } })
-    await prisma.enterprise.delete({ where: { id: supplierDeleted.id } })
-    await prisma.person.delete({ where: { id: supplierDeleted.id } })
+    if (supplierDeleted.enterpriseId) {
+      await prisma.enterprise.delete({ where: { id: supplierDeleted.id } })
+    }
+    if (supplierDeleted.personId) {
+      await prisma.person.delete({ where: { id: supplierDeleted.id } })
+    }
     await prisma.entity.delete({ where: { id: supplierDeleted.id } })
 
     res.status(201).json({ message: 'O fornecedor foi deletado.' })
